create CircleAnimation to make Circle.tsx DRYer
